feat: BlueMoonLadySajuResultPage 컴포넌트 추가 및 기존 페이지에서 사용
- 새로운 BlueMoonLadySajuResultPage 컴포넌트를 생성하여 웹툰 결과 페이지의 구조를 정의
- 기존 index.tsx에서 웹툰 관련 UI 요소를 BlueMoonLadySajuResultPage로 분리하여 코드 구조 개선
